refactor: 유틸클래스를 다시 생성한다
> [유틸 클래스 답변](https://github.com/next-step/kotlin-racingcar/pull/301#issuecomment-945790061)
