+++ src/main/kotlin/racingcar/utils/StringUtils.kt
@@ -1,0 +1,12 @@
+package racingcar.utils
+
+object StringUtils {
+
+    fun splitText(text: String, delimiter: String): List<String> {
+        return try {
+            text.split(delimiter)
+        } catch (e: IllegalArgumentException) {
+            emptyList()
+        }
+    }
+}
